+++ sdk/python/kfp/compiler/_read_write_test_config.py
@@ -58,6 +58,7 @@ CONFIG = {
             'concat_message',
             'dict_input',
             'identity',
+            'input_artifact',
             'nested_return',
             'output_metrics',
             'preprocess',
