+++ sdk/python/kfp/components/component_factory.py
@@ -17,7 +17,7 @@ import itertools
 import pathlib
 import re
 import textwrap
-from typing import Callable, List, Optional, Tuple, Union
+from typing import Callable, List, Optional, Tuple
 import warnings
 
 import docstring_parser
