+++ sdk/python/kfp/components/container_component_artifact_channel.py
@@ -1,9 +1,23 @@
+# Copyright 2022 The Kubeflow Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from typing import Union
 
 from kfp.components import placeholders
 
 
-class ContainerComponentArtifactChannel():
+class ContainerComponentArtifactChannel:
     """A class for passing in placeholders into container_component decorated
     function."""
